app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
requests==2.31.0
openai==1.52.0
jinja2==3.1.3
python-multipart==0.0.9
httpx==0.27.0
python-dotenv